Sequential day asssignment optimisation
Instead of each day in a period corresponding to a different prisoner, a single prisoner now corresponds to a series fo sequential days in a longer period. This means that visitors to the room later in the sequence for a prisoner will benefit from the combined knowledge of all previous visitors in that sequence; if any of them knew about the designated prisoner, then the later visitors will learn too.

Obviously this strategy has diminishing returns; spending 100 years on each prisoner's sequence will be less efficient than spending one day. By manual approximate optimisation, I decided that the optimal sequence per prisoner was approx. 40 days.

Changing the sequence length after 5k-20k days did not seem to improve the results.

The mean time-to-release is now ~150 days (down from ~250).